cleanup fit_bounds.js template
